A bit more advance version of Message Passing
The message passing now is composed this way. Each CPU has it's own output
"bucket" and input "bucket". When a message needs to be sent, it can fill it's
own output bucket and set a boolean flag ready. The server waiting for a message
will poll this ready "bucket" and when there is a change fetches the message
from the output bucket of the CPU.

Currently the ready bucket is just a boolean pointer, but in the future it
should be a matrix, so that each server can have it's own ready bucket.

These are the other small changes:

1. Remove an unused parameter in vmxon
2. Create a wait_ready function (this will later be in a library)
3. Created the msg_send and msg_receive functions (these will also be moved to a
   library soon)
4. Send a test message from the FS to the PM