Fixed a HUGE bug.
When EPT is mapped, the page tables were mapped just for 2 consecutive pages.
When these pages were in different addresses everything is messed up! Now that
we use dynamic allocation in PM this can happen (quite often) so we wrote a
function that maps all the page tables automatically wherever these are!