First test on batch processing.
With this patch I tested the possibility of sending multiple batch requests on
different fds but to the same driver.

I created a new driver called "junk" that always reply a constant value, and
created a new program called testcnl to test the possibility of sending
multiple requests to multiple fds of that driver.

NOTE: With this patch the minimum number of CPUs that needs to be available is
9, not 8 anymore.