Solve a Race Condition in the channels.
A race condition can happen between PM and the driver, since the channel list is
shared between the two CPUs and the PM changes it while the driver is reading
from it.
To solve the problem now there is a synchronization between the driver and the
PM/FS when a channel is closed from the program or if the program exists (and
the fds are all closed automatically).